server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
tweepy==4.14.0
aiohttp==3.9.1
python-multipart==0.0.6
